update the identification integration test
The identification-Isolated.sh script tests the identification module even when it isnt running. It requieres privilegies to run docker commands and at the end of the testing, the container is stopped.
On the other hand, the identification-FullAgent.sh script assumes that the full agent is in execution, including of course, the identification module.